requirements.txt
annotated-types==0.6.0
anyio==4.3.0
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
configobj==5.0.8
configparser==6.0.0
Deprecated==1.2.14
distro==1.9.0
etelemetry==0.3.1
exceptiongroup==1.2.0
filelock==3.13.1
google-ai-generativelanguage==0.4.0
google-api-core==2.17.1
google-auth==2.28.1
google-generativeai==0.3.2
googleapis-common-protos==1.62.0
grpcio==1.62.0
grpcio-status==1.62.0
h11==0.14.0
httpcore==1.0.4
httplib2==0.22.0
httpx==0.27.0
idna==3.6
img2pdf==0.5.1
isodate==0.6.1
looseversion==1.3.0
lxml==5.1.0
networkx==3.2.1
nibabel==5.2.0
nipype==1.8.6
numpy==1.26.4
openai==1.12.0
packaging==23.2
pandas==2.2.1
pathlib==1.0.1
pdf2jpg==1.1
pikepdf==8.13.0
pillow==10.2.0
proto-plus==1.23.0
protobuf==4.25.3
prov==2.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.6.2
pydantic_core==2.16.3
pydot==2.0.0
pyparsing==3.1.1
PyPDF2==3.0.1
pytesseract==0.3.10
python-dateutil==2.8.2
pytz==2024.1
pyxnat==1.6.2
rdflib==7.0.0
requests==2.31.0
rsa==4.9
scipy==1.12.0
simplejson==3.19.2
six==1.16.0
sniffio==1.3.0
tika==2.6.0
tqdm==4.66.2
traits==6.3.2
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.1
wrapt==1.16.0
